requirements.txt
httpx>=0.27
duckdb>=1.0
cryptography>=42.0
websockets>=12.0
numpy>=1.26
pandas>=2.1
scipy>=1.12
aiohttp>=3.9
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
jupyter>=1.0
matplotlib>=3.8
